+++ star_realms/src/game/components/card/details.rs
@@ -1,0 +1,7 @@
+
+type PlaySet = Vec<Play>;
+type Play = ...;
+
+pub enum Requirement {
+    Condition
+}
